+++ tvm-disasm-cli/src/main/kotlin/org/ton/disasm/Main.kt
@@ -20,6 +20,7 @@ import kotlinx.serialization.json.jsonObject
 import kotlinx.serialization.json.jsonPrimitive
 import org.ton.bytecode.TvmContractCode
 import org.ton.bytecode.disassembleBoc
+import org.ton.bytecode.prettyPrint
 import org.ton.net.TONCENTER_API_V3
 import org.ton.net.makeRequest
 import org.ton.net.toUrlAddress
@@ -103,7 +104,7 @@ class PrettyPrintDisassemblerCommand : CliktCommand(
         val contractCodeSource = contractCode
         val bocContent = fetchContractCode(contractCodeSource)
         val disassembledFile: TvmContractCode = disassembleBoc(bocContent)
-        val formattedStr = prettyPrint(disassembledFile, includeTvmCell)
+        val formattedStr = disassembledFile.prettyPrint(includeTvmCell)
         echo(formattedStr)
     }
 }
